Hope the best try
I've changed flto 1 and lr by 2 in 340 epoch
fl to 0 and pw to 16 at 350
After analyzing of the focal loss i can say that it's a problem with modulation factor. Focal loss does not split weights for positive and negative. Instead it use p or (1-p) for negative as p_t
I don't interested in focus on bad negative targets